Set ALLOWED_HOSTS to *
This is because when django is deployed with wsgi, if the host doesn't match ALLOWED_HOSTS, django raises a 500, with no explaination in the logs.
This took 24h to debug.
We think that limiting the allowed hosts like this adds nothing,
hence *.
Django REST Framework bypasses this security measure, so only the initial / -> /v1/ redirect was breaking, confusing the matter.

Doc:
https://docs.djangoproject.com/en/1.6/ref/settings/#allowed-hosts