Change the healthtest file, uploaded with "upload-healthtest.sh"
The Canonistack version of swift (at keystone.canonistack.canonical.com)
for some reason corrupts all text files when trying to download with
python swiftclient. You might upload a file with the contents "UP",
but when it is downloaded again, it'll have corrupted characters.

The problem doesn't apply to other installations of swift, so it should
be fine on Live, but for testing, we can't upload a "chinacachehealthtest.txt"
or "google-sdfgdsfgdsfgds.html" file without it being corrupted.

So for the healthtest for the deployment process, we've gone with a file
without an extension. For some reason files without extensions are fine.