get-token.sh to replace create-*.sh scripts
get-token.sh is a simpler to use version of the create-* scripts
It won't create another token if one exists, and it never lets you
specify the token yourself - I decided this is pointless.

It also doesn't prompt you for the name of the database.
If you're not using the default "assets", then you can just deal with
creating tokens manually yourself.