Upgrade djangorestframework to 3.0.5, update code
DjangoRestFramework 3 now provides the "get_indent" method
inside JSONRenderer, making our PrettyJsonRenderer simpler to
implement.

However, it also breaks the way we were using the
`AuthenticationFailed` exception, because the `detail`
argument is now forced to render as a single string
(rather than an array).

So I've written PrettyAuthenticationFailed to undo
this behaviour so we can still print a nice JSON array
to properly render the ASCII art.

For details see:
https://github.com/tomchristie/django-rest-framework/issues/2627
https://github.com/tomchristie/django-rest-framework/issues/2341
https://gist.github.com/nottrobin/5b7fc7fd8ae08c2b1ea3
https://gist.github.com/nottrobin/e06af63f4fb3aee56bb7