Handle parameters more gracefully
If any parameters were passed to an image that Pilbox didn't understand,
we were getting an internet server error.

We've fixed this by only passing recognised "resize" parameters to Pilbox
(if no "op" setting is supplied)

If you do provide an "op" setting without other parameters, so that Pilbox throws an error
we now provide a nice friendly "400" message.