url_helpers, containing "normalize" helper
The "normalize" function will make sure strings are always encoded
such that spaces will become "+". This is so that we can make sure
in all places in the assets server, filenames are stored with "+"
instead of spaces. This makes everything much easier, since we don't
have to worry so much about URL encoding.