URL paths for managing redirects
Make use of the new RedirectRecord and RedirectRecords
view classes, and the new RedirectMapper to allow for
manipulation of redirects.

These redirects are stored in the "redirects" collection
in MongoDB, and have a "request_path" and a "target_url".

You can manipulate them with the following URLs and HTTP methods:

/redirects - GET, POST
/redirects/<request_path> - GET, PUT, DELETE