Remove upload-healthtest.sh
upload-healthtest.sh is run during deployment to create a healthtest
asset, which we can then use to check the assets-server is serving
assets correctly.

This is a function we still need, but I don't think the code to do it
should be stored within the codebase, as it's only relevant to our
specific deployment setup.

Therefore, we should instead be uploading the healthtest file through the
server's API with `curl` or similar from the deployment environment.
This would also be a better integration test, as we'd be testing the
standard API interface at the same time.