Unify and simplify integration test
The integration test should ask for the server URL and token directly.
It has no need to store a fixture.

Uptimately, this test probably shouldn't live in this codebase, since it
doesn't actually need to link into the codebase in any way. But for
now it's probably the simplest place to keep it.