Look for mongo on localhost by default
Instead of expecting a network alias of `db` to be available, we should
default to looking on `localhost` in the settings. This is because, if
someone was coming to the project fresh without knowledge of our `./run`
infrasturcture, they would expect to simply get the local server running
with a local database.

So instead, we should explicitly state where to find the database in the
`./run` script, using `DATABASE_URL`.